requirements.txt
web.py
nose
requests
PyYAML
jsonpath-rw
pydash